requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
boto3==1.34.153
botocore==1.34.153
certifi==2024.7.4
channels==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Django==4.1
django-cors-headers==4.4.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.96
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.23.0
numpy==1.26.4
openai==1.38.0
openpyxl==3.1.5
orjson==3.10.6
packaging==24.1
pandas==2.2.2
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.1
tablib==3.6.1
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
Django==4.1
djangorestframework
django-cors-headers
pandas
tablib
channels
openpyxl
psycopg2
openai
pymupdf
requests
celery
scikit-learn
nltk
transformers
pdfplumber